+++ tests/test_axisgrid.py
@@ -8,7 +8,7 @@ import numpy.testing as npt
 from numpy.testing import assert_array_equal, assert_array_almost_equal
 import pandas.testing as tm
 
-from seaborn._oldcore import categorical_order
+from seaborn._base import categorical_order
 from seaborn import rcmod
 from seaborn.palettes import color_palette
 from seaborn.relational import scatterplot
