+++ tests/test_relational.py
@@ -10,7 +10,7 @@ import pytest
 from numpy.testing import assert_array_equal, assert_array_almost_equal
 
 from seaborn.palettes import color_palette
-from seaborn._oldcore import categorical_order
+from seaborn._base import categorical_order
 
 from seaborn.relational import (
     _RelationalPlotter,
